vector: slightly improve noexcept backing resize
This avoids a "throw will always call terminate" warning when compiling
with exceptions enabled. The compiler (GCC 11.4) isn't smart enough to
realise that the catch block won't execute if the call within the try
block is noexcept and so issues a warning due to rethrow inside a
function (__enlarge_capacity) that's then also noexcept.