swap: fix to work with clang
swap is required not to participate in overload resolution in some
circumstances, which was implemented by adding a second template
parameter, defaulting to an expression that would SFINAE-out if
applicable.

Both the definition and forward declaration (in type_traits) included
the default type expression for the enabling parameter, but it turns out
this was bad C++ and clang rejects it as does GCC 12+. The default can
not be specified twice.

Fix by removing the 2nd template parameter, and using enable_if_t<> for
the return type instead.